antiword half fixed

+++ auto/antiword.js
@@ -1,6 +1,6 @@
 client.on("MessageCreate", message =>{
 const wordz = ["sus", "amongus", "bald", "bozo"]; 
-if (wordz.some(word => message.content.toLowerCase().includes(word.toLowerCase()))) { 
+if (wordz.some(word => message.content.toLowerCase().includes(word.toLowerCase())) && message./*make it permission(admin) and owner sensitive*/) { 
   message.reply("bad word bad");
   delay(2500);
   message.delete();
